pyproject.toml
[tool.poetry]
name = "awsome-backend"
version = "0.1.0"
description = ""
authors = ["Hemant Kumar A <hemant.ahk04@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
loguru = "^0.7.2"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.25"
asyncpg = "^0.29.0"
pytz = "^2024.1"
slowapi = "^0.1.9"
aiosql = "^9.4"
pypika = "^0.48.9"
asgi-lifespan = "==2.*"
alembic = "1.12.1"
psycopg2-binary = "^2.9.9"

[tool.poetry.dev-dependencies]
black = "^24.1.1"
isort = "^5.13.2"
pre-commit = "^3.6.0"
pytest-cov = "^4.1.0"
pytest = "^8.0.0"
bump-pydantic = "^0.8.0"
anyio = "^4.2.0"
pytest-asyncio = "^0.23.5"
pytest-tornasync = "^0.6.0.post2"
pytest-trio = "^0.8.0"
httpx = "^0.26.0"
pytest-xdist = "^3.5.0"
pytest-env = "^1.1.3"






[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
